Added intersection observer, changed the hover animations
Added an observer to see which section the user is in, indicating it by underlining the respective link in the navigation bar. Hover animations were also changed, creating a class that will fire when the user hovers over the link and when the view is inside a specific section (this was done precisely to make it easier for the observer to code)

The html scroll behavior was changed to smooth in order to avoid delay when the view is redirected to any section.